test: 100% test coverage :tada:
Make CLI, which is mainly for testing, use the no-argument form of
Logger() when possible in order to achieve 100% branch coverage in
Logger() function.

Remove the un-used support for boolean flags that default to true (would
emit a --no-OPT entry in the usage message), bringing the cli itself up
to 100% coverage.

This package now has 100% line and branch coverage.

Signed-off-by: Ryan Graham <r.m.graham@gmail.com>