src: replace byline with custom splitter
Rather than continue depending on an external module with its own
dependencies, not all of which are still maintained, replace our usage
of it with a minimal implementation of a line splitting Transform
stream.

In addition to the StringDecoder related tests, add a blank line to one
of the test examples so that we can make sure we enforce the current
behaviour of swallowing empty lines.

This implementation follows the same general patter as byliner, which is
to use a StringDecoder to accumulate bytes so we can ensure that we're
not assuming everything is single byte characters and accidentally
interpreting a random byte in the middle of a multi-byte sequence as
something it isn't and then corrupting things as a result.

For anyone looking to level up their use of streams, this is the kind of
thing most people don't think of and an approach you should absolutely
consider when dealing with messages on streams that may be split across
reads such as splitting a multi-byte UTF8 character across a packet
boundary on a TCP request.

Signed-off-by: Ryan Graham <r.m.graham@gmail.com>