src: replace byline with custom splitter
The byline package is generating deprecation warnings when running
strong-log-transformer under node 10. Since the package appears to be
unmaintained, we are re-implementing the functionality here.

The implementation is heavily inspired by byline, split, and split2, but
simplified for our specific use case.

The existing tests fully cover this new implementation, including the
corner case of the input stream being truncated in the middle of a
multi-byte UTF8 sequence.

Resolves #7.

Signed-off-by: Ryan Graham <r.m.graham@gmail.com>