src: replace through with readable-stream
The convenience API provided by the through module is no longer worth
the cost of the dependency because the API provided by streams in core
is equally convenient for our uses.

Additionally, by switching to using readable-stream, which is stream
from node core published as a module, we can ensure our use of the
stream API will work across multiple versions of Node, even if they are
too old to include certain features in their internal stream
implementation.

Signed-off-by: Ryan Graham <r.m.graham@gmail.com>