test: add tests for truncated utf8 sequences
If a source gives us an invalid utf8 sequence, ensure that it is
replaced by a Unicode replacement character as expected.

One of the semver-major changes in node 8 was to change string_decoder
behaviour to better align with the implementation in v8 itself. Now that
we have tests for the handling of incomplete utf8 sequences, this
actually affects us. To account for this, use a simple version check to
determine the expected string output of our bad-utf8 tests.

For background see nodejs/node@24ef1e6775

Signed-off-by: Ryan Graham <r.m.graham@gmail.com>