test: add color tag tests
The majority of the uses of strong-log-transformer are for prefixing
console output from multiple processes with a child process identifier
and almost all of them make use of ANSI escape codes in the tags so that
the output is suitably colorful. Since that's a common use, these tests
will ensure that the current behaviour doesn't accidentally change.

Signed-off-by: Ryan Graham <r.m.graham@gmail.com>